added utils file
